Extract generic brace range partial
